requirements.txt
Flask
PyMuPDF
rapidfuzz
google-generativeai
